### 2022-11-18 #4950 @editorialbot
[PRE REVIEW]: discourseGT: An R package to analyze discourse networks in educational contexts
**Submitting author:** @q1cui (<a href="http://orcid.org/0000-0002-3034-1143">Qi Cui</a>)
**Repository:** https://github.com/q1cui/discourseGT
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.8
