### 2022-11-28 #4950 @q1cui
> Hi, @q1cui, I will be the editor of your submission.
>
> It seems that we cannot generate a manuscript proof. Can you check if your paper.md file follows the format shown in this example: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography ?
>
> At the same time, we can start looking for reviewers for you submission. We usually ask the authors for suggestions, so let me know if you have anyone in mind that could do this - you can paste the github profile of anyone you think that could do it (but please don't tag anyone yet). You can also check [this list of people](https://bit.ly/joss-reviewers) that have agreed to review for JOSS, and see if any of them may be suitable to review this submission.

Thanks, @marcosvital We have checked our paper.md file and removed the invalid ORCID. We find zoometh, OwenWard, gabett, schochastics, chrisaberson are suitable to review our submission.
