### 2023-01-09 #4950 @q1cui
> Hi, @q1cui, it seems that you reference file is missing (or named differently), as we got this error message:
>
> > PandocResourceNotFound "references.bib"

Thank you for your reply! I put reference.bib in a different folder, but now it is added to the paper folder. Thanks!
