### 2023-01-09 #4950 @marcosvital
OK, it's working now! @q1cui please check the manuscript to see if it's ready for review. If you do any changes, you can call the editorialbot to generate a new pdf any time, ok?
