+++ vignettes/references.bib
@@ -728,3 +728,15 @@ title = {Analysing students interaction process for improving collaboration. The
 volume = {11},
 journal = {JAIED}
 }
+
+@article{Lou_2001,
+author = {Lou, Y., Abrami, P. C., & d’Apollonia, S.},
+title = {Small group and individual learning with technology: A meta-analysis},
+journal = {Review of Educational Research},
+volume = {71},
+number = {3},
+pages = {449-521},
+year = {2001},
+doi = {10.3102/00346543071003449},
+URL = {https://journals.sagepub.com/doi/10.3102/00346543071003449}
+}
