### 2023-01-09 #4950 @marcosvital
> > OK, it's working now! @q1cui please check the manuscript to see if it's ready for review. If you do any changes, you can call the editorialbot to generate a new pdf any time, ok?
>
> Thanks a lot! I realized there might be a problem with one reference (Lou 2001) format, I just updated it, other than that I think it's ready for review. Thank you very much for your help!

OK, that's great! Once we get the reviewers we'll move on to reviewing phase.
