### 2023-01-27 #4950 @q1cui
Dear Dr. Vital,

Since not all suggested reviewers are able to review the manuscript, I was wondering if need to suggest more reviewers. Please feel free to let me know what else I can do on my end.

Thanks,
Qi

Qi Cui, Ph.D.
Postdoctoral Scholar
Section of Cell and Developmental Biology
Division of Biological Sciences
University of California, San Diego
Email: ***@***.***

From: Marcos Vital ***@***.***>
Sent: Monday, January 9, 2023 1:46 PM
To: ***@***.***>
Cc: Cui, ***@***.***>; ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: discourseGT: An R package to analyze discourse networks in educational contexts (Issue #4950)

OK, it's working now! @q1cui<https://urldefense.com/v3/__https:/github.com/q1cui__;!!Mih3wA!ApZS1-yzkcQEC9awcslrxvek2V1GQZABDDxHjNEovnS82Vl5WrrX8uShERafPAiyDva81aN5nBrU-uzIO4HNJPPu$> please check the manuscript to see if it's ready for review. If you do any changes, you can call the editorialbot to generate a new pdf any time, ok?

Thanks a lot! I realized there might be a problem with one reference (Lou 2001) format, I just updated it, other than that I think it's ready for review. Thank you very much for your help!

OK, that's great! Once we get the reviewers we'll move on to reviewing phase.
