### 2023-02-09 #4950 @lebebr01
Thanks for thinking of me. I should be able to do a review by the end of
February if that timeline works for you.

Best
Brandon

On Thu, Feb 9, 2023, 6:33 AM Marcos Vital (LEQ-UFAL) <
***@***.***> wrote:

> Dear @lebebr01 <https://github.com/lebebr01>, would you be willing to
> review this submission for JOSS?
>
> We carry out an open checklist-driven reviews here in GitHub issues, and
> follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> This is still a pre-review issue, and once there are enough reviewers we
> will move on to a new issue where the review will take place. If you are
> not able to review, please let us know if you have recommendations for
> other reviewers that could help with this submission.
>
