### 2023-02-10 #4950 @marcosvital
Thank you for your quick reply, @lebebr01.

Yes, it should be fine if you start by the end of the month. The others reviewers will start soon if they are available to, and you will be able to follow the discussion later, since the reviewing process is open.
