### 2023-02-10 q1cui/discourseGT#1 @schochastics
paper not reproducible
The first lines of code in the paper are
```r
# Calculate the weighted edge list
tabEdge <- tabulate_edges(data, iscsvfile = FALSE, silentNodes = 0)
# Check the weighted edge list
head(tabEdge$master)
```
where data is undefined.
Please add and rename
```r
install.packages("discourseGT")
library(discourseGT)
data(sampleData1)
# Calculate the weighted edge list
tabEdge <- tabulate_edges(sampleData1, iscsvfile = FALSE, silentNodes = 0)
# Check the weighted edge list
head(tabEdge$master)
```
so data should be replaced with sampleData1 throughout the paper

This is part of the review for JOSS https://github.com/openjournals/joss-reviews/issues/5143
