### 2023-02-10 q1cui/discourseGT#2 @schochastics
function output problems
`summaryNet()`

Don display the whole
` ---------DISCLAIMER AND WARRANTY OF PROVIDED RESULTS AND CODE---------`
part. You can put that into the code as a comment

`plotNGTData`

There are a whole lot of warnings when running the function
```r
#>$episodes_plot
#> Use of `eps_st$episode_length` is discouraged.
#>ℹ Use `episode_length` instead.
#>Warning: Use of `eps_st$Freq` is discouraged.
#>ℹ Use `Freq` instead.
#> $ntr_plot
#> Warning: Use of `temp_plot$participant` is discouraged.
#> ℹ Use `participant` instead.
#> Warning: Use of `temp_plot$normalized_turn_ratio` is discouraged.
#> ℹ Use `normalized_turn_ratio` instead.
#> Warning: Use of `temp_plot$group` is discouraged.
#> ℹ Use `group` instead.
#> Use of `temp_plot$group` is discouraged.
#> ℹ Use `group` instead.
#> Warning: Use of `temp_plot$participant` is discouraged.
#> ℹ Use `participant` instead.
#> Warning: Use of `temp_plot$normalized_turn_ratio` is discouraged.
#> ℹ Use `normalized_turn_ratio` instead.
#> Warning: Use of `temp_plot$group` is discouraged.
#> ℹ Use `group` instead.
#> Use of `temp_plot$group` is discouraged.
#> ℹ Use `group` instead.
```

This is part of the review for JOSS https://github.com/openjournals/joss-reviews/issues/5143
