### 2023-02-10 q1cui/discourseGT#3 @schochastics
add tests
[Add some unit tests](https://r-pkgs.org/testing-basics.html). You do note need to have 100% coverage but a few tests would be nice to have

This is part of the review for JOSS https://github.com/openjournals/joss-reviews/issues/5143
