### 2023-02-10 q1cui/discourseGT#4 @schochastics
add info on contributing to the package
Here is a nice example: https://github.com/cosimameyer/overviewR#how-to-reach-out

This is part of the review for JOSS https://github.com/openjournals/joss-reviews/issues/5143
