### 2023-02-13 q1cui/discourseGT#5 @schochastics
improve pkg docs
?discourseGT is too short. A description of the key functions would be helpful to have here

This is part of the review for JOSS https://github.com/openjournals/joss-reviews/issues/5143
