### 2023-02-13 q1cui/discourseGT#6 @schochastics
function parameters
The use of `iscsvfile = TRUE` to indicate that the first parameter is supposed to be interpreted as  path is a bit redundant.
Better check in the function if the input is a character (hence a path to a file) or a data frame.

This is not a mandatory fix for the review, just a suggestion to make the package follow good practices

This is part of the review for JOSS https://github.com/openjournals/joss-reviews/issues/5143
