### 2023-02-13 #5143 @schochastics
I have added a few issues to the repository which I would like to authors to address. Especially the test part is crucial to me. Please take some time and effort to add some tests to the package.

Overall this is a very nice package which just lacks a bit of good practice but that should be fixable. I am happy with the submission once my issues are addressed adequately. (cc @q1cui)
