+++ paper/references.bib
@@ -90,7 +90,7 @@ number = {4},
 pages = {361-428},
 year = {2003},
 publisher = {Routledge},
-doi = {10.1207/s1532690xci2104\_2},
+doi = {10.1207/s1532690xci2104_2},
 URL = {https://doi.org/10.1207/s1532690xci2104_2},
 eprint = {https://doi.org/10.1207/s1532690xci2104_2}
 }
@@ -138,7 +138,7 @@ number = {2},
 pages = {101-111},
 year = {2000},
 publisher = {Routledge},
-doi = {10.1207/S15326985EP3502\_4},
+doi = {10.1207/S15326985EP3502_4},
 URL = {https://doi.org/10.1207/S15326985EP3502_4},
 eprint = {https://doi.org/10.1207/S15326985EP3502_4}
 }
@@ -153,7 +153,8 @@ pages = {599--623},
 publisher = {Ohio State University Press},
 title = {Classrooms as Communities: Exploring the Educational Character of Student Persistence},
 volume = {68},
-year = {1997}
+year = {1997},
+doi = {10.1080/00221546.1997.11779003},
 }
 
 @article{Loes_2017,
@@ -255,7 +256,8 @@ URL = {https://www.springer.com/gp/book/9780387952413},
 author = {Chris Godsil and Gordon F. Royle},
 publisher = {Springer-Verlag},
 title = {Algebraic Graph Theory},
-year = {2001}
+year = {2001},
+doi = {10.1007/978-1-4613-0163-9}
 }
 
 @article{Girvan_2002,
@@ -288,7 +290,7 @@ number = {1},
 pages = {7-15},
 year = {1989},
 issn = {0020-0190},
-doi = {https://doi.org/10.1016/0020-0190(89)90102-6},
+doi = {10.1016/0020-0190(89)90102-6},
 url = {http://www.sciencedirect.com/science/article/pii/0020019089901026},
 author = {Tomihisa Kamada and Satoru Kawai}
 }
@@ -299,7 +301,8 @@ author = {Edward M. Reingold and John S. Tilford},
 journal = {IEEE Transactions on Software Engineering},
 volume = {7},
 number = {2},
-year = {1981}
+year = {1981},
+doi = {10.1109/TSE.1981.234519}
 }
 
 @manual{R-GGally,
@@ -326,6 +329,7 @@ volume = {13},
 pages = {2498-2504},
 year = {2003},
 url = {https://www.cytoscape.org/},
+doi = {10.1101/gr.1239303}
 }
 
 @manual{R-rcy3,
@@ -393,6 +397,7 @@ author = {Kevin C. Dunn and Iver B. Neumann},
 publisher = {University of Michigan Press},
 title = {Undertaking Discourse Analysis for Social Research},
 year = {2016},
+doi = {10.3998/mpub.7106945}
 }
 
 @article{Liyanage_2021,
@@ -428,7 +433,7 @@ journal = {Bell System Technical Journal},
 volume = {27},
 number = {3},
 pages = {379-423},
-doi = {https://doi.org/10.1002/j.1538-7305.1948.tb01338.x},
+doi = {10.1002/j.1538-7305.1948.tb01338.x},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/j.1538-7305.1948.tb01338.x},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/j.1538-7305.1948.tb01338.x},
 year = {1948},
@@ -441,7 +446,7 @@ volume = {13},
 pages = {131-144},
 year = {1966},
 issn = {0022-5193},
-doi = {https://doi.org/10.1016/0022-5193(66)90013-0},
[…]
