### 2023-02-26 #5143 @zoometh
👋 @q1cui

Some minor errors:

* missing ORCID | typos:

![image](https://user-images.githubusercontent.com/67924910/221434347-9c626739-55ce-4d18-bec7-f0a8589745b2.png)

* one missing reference:

![image](https://user-images.githubusercontent.com/67924910/221434555-ab169fb8-5e35-42bc-93ac-d13e24b473b5.png)

* change `null` to uppercase (`NULL`) since R is case sensitive:

![image](https://user-images.githubusercontent.com/67924910/221434824-7157a287-7e6c-4312-9e9a-e574c8facf28.png)

* No clear Community guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. Please add them in GH
