### 2023-03-07 de84b7a
Update paper.md
data is undefined, it is replaced with sampleData1
