### 2023-03-07 5d3acc9
Update plotNGTData.R
fix some warnings:
#>$episodes_plot
#> Use of `eps_st$episode_length` is discouraged.    (line105)
#>ℹ Use `episode_length` instead.
#>Warning: Use of `eps_st$Freq` is discouraged.
#>ℹ Use `Freq` instead.
#> $ntr_plot
#> Warning: Use of `temp_plot$participant` is discouraged.    (line315)
#> ℹ Use `participant` instead.
#> Warning: Use of `temp_plot$normalized_turn_ratio` is discouraged.
#> ℹ Use `normalized_turn_ratio` instead.
#> Warning: Use of `temp_plot$group` is discouraged.
#> ℹ Use `group` instead.
#> Use of `temp_plot$group` is discouraged.
#> ℹ Use `group` instea
[…]
