+++ vignettes/discourseGT.Rmd
@@ -3,39 +3,10 @@ title:
   formatted:  "\\pkg{discourseGT}: An \\proglang{R} package to analyze discourse networks in educational contexts"
   plain:     "discourseGT: An R package to analyze discourse networks in educational contexts"
   short:     "\\pkg{discourseGT}: An \\proglang{R} package to analyze educational discourse networks"
-documentclass: jss
-citation_package: natbib
-output: rticles::jss_article
-classoption: nojss
-bibliography: references.bib
-bibliography_style: plainnat
 vignette: >
   %\VignetteIndexEntry{discourseGT}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
-author: 
-  - name: Joshua P. Le
-    affiliation: 'University of California `\\`{=latex} San Diego'
-  - name: Albert Chai
-    affiliation: 'University of California `\\`{=latex} San Diego \AND'
-  - name: Andrew S. Lee
-    affiliation: 'University of California `\\`{=latex} Los Angeles'
-  - name: Kevin Banh
-    affiliation: 'University of California `\\`{=latex} San Diego \AND'
-  - name: Priya Pahal
-    affiliation: 'University of California `\\`{=latex} San Diego'
-  - name: Stanley M. Lo
-    affiliation: 'University of California `\\`{=latex} San Diego'
-    address: >
-      9500 Gilman Dr #0355
-      La Jolla, CA 92093-0355
-    email: \email{smlo@ucsd.edu}
-abstract: >
-  Student collaborations and discussions in the classroom are important components of the learning process. Current methods for analyzing discourse in these educational contexts are predominantly qualitative. To complement the existing methodologies in education research, \pkg{discourseGT} is an \proglang{R} package that adapts graph theory and social network analysis to analyze conversation patterns when students collaborate in small groups. Ths software package takes data on the sequential student talk-turns in a classroom discussion and produces statistics and plots based on both graph theory and other non-graph theory parameters. Overall, these new features in \pkg{discourseGT} can provide insight on the dynamics of student discussions relevant to education researchers.
-
-keywords: 
-  formatted:  [student group work, discourse network, graph theory, "\\pkg{discourseGT}", "\\proglang{R}"]
-  plain:      [student group work, discourse network, graph theory, discourseGT, R]
 preamble: >
   \usepackage{longtable}
   \usepackage{makecell}
@@ -57,20 +28,6 @@ graphics: yes
 ```{r, include = FALSE}
 options(prompt = "R> ", continue = "+ ", width = 70, useFancyQuotes = FALSE)
 ```
-
-# Introduction
-Vygotsky postulated that social interactions are crucial to individual learning and cognitive development [@Vygotsky_1978]. Subsequent studies in education research have found that small-group discussions can help students develop disciplinary understanding [@Freeman_2014] and enhance skills in problem solving [@Heller_1992], critical thinking [@Webb_1982b, @Gokhale_1995, @Bligh_2000], communication [@Webb_1994], and metacognition [@Webb_2003, @Veenman_2006, @Bromme_2010]. Furthermore, small-group discussions can also improve student interest and motivation [@Skinner_1993, @Ryan_2000] as well as course completion rates and persistence in science, technology, engineering, and mathematics (STEM) majors [@Tinto_1997, @Freeman_2014, @Loes_2017].
-
[…]
