### 2023-03-11 d4ef1cf
Update README.md
add Community guidelines: clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
