### 2023-03-15 cc592b4
Delete discourseGT.Rmd
