+++ vignettes/discourseGT.Rmd
@@ -1,339 +1,0 @@
----
-title:
-  formatted:  "\\pkg{discourseGT}: An \\proglang{R} package to analyze discourse networks in educational contexts"
-  plain:     "discourseGT: An R package to analyze discourse networks in educational contexts"
-  short:     "\\pkg{discourseGT}: An \\proglang{R} package to analyze educational discourse networks"
-vignette: >
-  %\VignetteIndexEntry{discourseGT}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
-preamble: >
-  \usepackage{longtable}
-  \usepackage{makecell}
-  \usepackage{tabularx}
-  \usepackage{hyperref}
-  \usepackage{graphicx}
-  \usepackage{amsmath}
-  \usepackage{booktabs}
-  \usepackage{amsfonts}
-  \usepackage{tabulary}
-  \usepackage{ragged2e}
-  \usepackage{floatrow}
-  \floatsetup[table]{capposition=top}
-  \floatsetup[figure]{capposition=top}
-longtable: true
-graphics: yes
----
-
-```{r, include = FALSE}
-options(prompt = "R> ", continue = "+ ", width = 70, useFancyQuotes = FALSE)
-```
-
-## Graph Theory Parameters
-In our previous work [@Chai_2019, @Liyanage_2021], we identified a subset of graph theory parameters that are relevant to small-group discussions in educational contexts as shown in Table \ref{table:gt_param}.
-
-\begin{table}[H]
-\setlength{\tymin}{70pt}
-\let\raggedright\RaggedRight
-\begin{tabulary}{\textwidth}{@{}LLLL@{}}
-\toprule
-\textbf{Parameter}      & \textbf{Graph Theory \newline Definition}                                  & \textbf{Social Network Definition}                         & \textbf{Discourse Network Definition}                          \\ \midrule
-Node           & An object of interest                                    & Typically a person                                & Participant in a small group discussion               \\
-Edge           & A connector between two nodes                            & Flow of information between two people            & Talk-turn between two individuals                     \\
-Direction      & Defines which node points to another using the edge      & Indicates which person has ties to the other      & Indicates which individual talks after the other      \\
-Weight         & A number associated with an edge                         & Frequency of information flow between two people  & Frequency of a talk-turn between two individuals      \\
-Degree         & Number of edges connected to a node                      & Number of people an individual has ties to        & Number of people an individual talks before and after \\
-Density        & Number of edges divided by the number of possible edges  & Interactions occurring among different people     & Talk-turns occurring among different individuals      \\
-Centrality     & A number for the importance of a given node in the graph & Amount of influence of each person                & Amount of talk-turn contribution of an individual     \\
-Centralization & A number for the importance of the central node          & Dependence of a network on its most active person & Dependence of a group on its most active individual   \\
-Subgraph       & A smaller graph within a graph                           & Individuals who have closer ties to each other    & Individuals who talk after each other more            \\ \bottomrule
-\end{tabulary}
-\caption{Graph theory parameters used in \pkg{discourseGT}}
-\label{table:gt_param}
-\end{table}
-
-## Non-Graph Theory Parameters
[…]
