### 2023-03-15 9667be8
Update README.md
adding key functions and examples document
