+++ README.md
@@ -12,11 +12,11 @@ Graphs, nodes, and edges may be characterized in various ways. For example, the 
 To get started installing the package, run the command `install.packages('discourseGT')` in R, which will download the most stable package from CRAN.  
 
 ## How to Use
-The key functions and examples of how to use this package can be found in ![this (https://github.com/q1cui/discourseGT/blob/main/vignettes/discourseGT_new.pdf)] document
+The key functions and examples of how to use this package can be found in [this document](https://github.com/q1cui/discourseGT/blob/cc592b48f9e0e70bf6ae82c91b6008b81b3f94e9/vignettes/discourseGT_new.pdf).
 
 ## How to reach out?
 ### Where do I report bugs?
-Simply open an issue on GitHub.
+Please open an issue on GitHub.
 
 ### How do I contribute to the package?
 If you have an idea (but no code yet), open an issue on GitHub. If you want to contribute with a specific feature and have the code ready, fork the repository, add your code, and create a pull request.
