### 2023-03-15 q1cui/discourseGT#5 @q1cui
Thanks for pointing it out. We have added a description of the workflow, key functions, and examples. If you click the hyperlink under _how to use_ in README.md, you could see the descriptions.
