### 2023-03-15 q1cui/discourseGT#5 @schochastics
Thank you. could you link this in the file below
https://github.com/q1cui/discourseGT/blob/d0b1fbbc2230579797652a9fb3afd8b0ea703903/R/discourseGT.R#L1

As a reference, see for instance
```r
library(dplyr)
?dplyr
```
