### 2023-05-12 cc5cd71
Update discourseGT.R
add key functions to R documentation
