### 2023-05-12 q1cui/discourseGT#2 @q1cui
We have fixed plotNGTData(), it is free of warnings in Github.
We also fixed summaryNet(), ----DISCLAIMER AND WARRANTY OF PROVIDED RESULTS AND CODE---- is not included in the output.
