### 2023-05-12 q1cui/discourseGT#2 @q1cui
We tested these functions by devtools::install_github("q1cui/discourseGT") for now.
