### 2023-05-12 q1cui/discourseGT#5 @q1cui
Thank you for your comments. We have added key function descriptions to this file: discourseGT/R/discourseGR.R
