### 2023-05-12 q1cui/discourseGT#6 @q1cui
The parameter `iscsvfile = TRUE` is used to help differentiate the two possible input types for `input_file` as you mentioned: path to CSV file(character) or a loaded R data frame. It is used as a protection to ensure that users are aware of the input being passed into the function.

My understanding of your suggestion is to check the input type within the function itself, a sample modification is copied below, please correct me if this is not what you were suggesting.

edgelist_raw <- function(input_file) {

  # check input type, read external input file into data.frame
  if (is.data.frame(input_file)) {
    df <- input_file
  } else if (is.character(input_file)) {
    df <- read.csv(input_file)
  } else {
    warning("Invalid input type. Please input file path as character or input data.frame directly.")
  }

  \# rest of the code
  \```
}
