+++ testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(discourseGT)
-
-test_check("discourseGT")
