### 2023-05-13 2ce64df
Update test_coreNetAnalysis.R
