### 2023-05-13 6625546
Update test_coreNetAnalysis.R
