### 2023-05-14 #5143 @q1cui
I greatly appreciate the constructive feedback @schochastics provided on my manuscript. Your insights are valuable, and they have given me a broader perspective on how to improve the manuscript. Here are the responses to each of your comments:

1. The use of iscsvfile = TRUE to interpret the first parameter as a path might be a bit redundant.
The parameter iscsvfile = TRUE is used to help differentiate the two possible input types for input_file as you mentioned: path to CSV file(character) or a loaded R data frame. It is used as a protection to ensure that users are aware of the input being passed into the function.
However, taking your suggestion into consideration, I understand that checking the input type within the function itself might be a more efficient and good practice. If there is any misinterpretation of your suggestion, please correct me. Here is a sample modification:
edgelist_raw <- function(input_file) {
#check input type, read external input file into data.frame
if (is.data.frame(input_file)) {
df <- input_file
} else if (is.character(input_file)) {
df <- read.csv(input_file)
} else {
warning("Invalid input type. Please input file path as character or input data.frame directly.")
}

# rest of the code
```
}

2. It was noted that the ?discourseGT is too short and could benefit from a description of the key functions.
Thank you for pointing this out. To address this, we have expanded the ?discourseGT section to include descriptions of the key functions. These updates can be found in this file: discourseGT/R/discourseGT.R.

3. The reviewer identified issues with plotNGTData() and summaryNet() functions.
We have addressed these concerns and ensured that plotNGTData() is free of warnings on Github. Additionally, we have corrected summaryNet() so that the ----DISCLAIMER AND WARRANTY OF PROVIDED RESULTS AND CODE---- is no longer included in the output.

4. It was suggested that the addition of some unit tests would be beneficial.
In response to this suggestion, we have introduced a few unit tests that can be found in the tests folder. We appreciate this suggestion and will continue to develop and expand these tests in the future.

I hope these revisions and clarifications are in line with your expectations. Once again, I appreciate your constructive feedback and look forward to your continued guidance.
