### 2023-05-15 q1cui/discourseGT#7 @schochastics
vignette build error
I am getting the following error when building the vignette
```r
R> devtools::check()
══ Documenting ═══════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
ℹ Installed roxygen2 version (7.2.3) doesn't match required (7.0.2)
✖ `check()` will not re-document this package
══ Building ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
Setting env vars:
• CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
• CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
• CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX14FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX17FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX20FLAGS: -Wall -pedantic -fdiagnostics-color=always
── R CMD build ───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/home/david/Documents/github_projects/discourseGT/DESCRIPTION’ ...
─  preparing ‘discourseGT’:
E  creating vignettes (12.6s)
   --- re-building ‘discourseGT_new.Rmd’ using rmarkdown
   Warning: Package tabulary Warning: No suitable columns! on input line 213.
   --- finished re-building ‘discourseGT_new.Rmd’

   --- re-building ‘discourseGT.pdf.asis’ using asis
   Error: processing vignette 'discourseGT.pdf.asis' failed with diagnostics:
   No such output file: discourseGT.pdf
   --- failed re-building ‘discourseGT.pdf.asis’

   SUMMARY: processing the following file failed:
     ‘discourseGT.pdf.asis’

   Error: Vignette re-building failed.
   Execution halted
Error in `(function (command = NULL, args = character(), error_on_status = TRUE, …`:
! System command 'R' failed
---
Exit status: 1
stdout & stderr: <printed>
---
```

This is part of a [https://github.com/openjournals/joss-reviews/issues/5143](JOSS Review)
