### 2023-05-15 q1cui/discourseGT#3 @schochastics
Several tests are failing

```r
R> devtools::test()
ℹ Testing discourseGT
Registered S3 method overwritten by 'GGally':
  method from
  +.gg   ggplot2
✔ | F W S  OK | Context
⠹ |         3 | basicPlot                                                                                                                                         ✖ | 1       5 | basicPlot [1.4s]
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Failure (test_basicPlot.R:27): basicPlot works
`basicPlot(...)` did not produce any warnings.
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✖ | 1       0 | coreNetAnalysis
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test_coreNetAnalysis.R:14): (code run outside of `test_that()`)
Error in `data.frame(master)`: object 'master' not found
Backtrace:
 1. discourseGT::prepareGraphs(raw_data_input, project_title)
      at test_coreNetAnalysis.R:14:0
 3. base::with.default(raw_data_input, data.frame(master))
 4. base::eval(substitute(expr), data, enclos = parent.frame())
 5. base::eval(substitute(expr), data, enclos = parent.frame())
 6. base::data.frame(master)
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✖ | 1       0 | plot1Att
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test_plot1Att.R:21): (code run outside of `test_that()`)
Error in `data.frame(master)`: object 'master' not found
Backtrace:
 1. discourseGT::prepareGraphs(raw_data_input, project_title)
      at test_plot1Att.R:21:0
 3. base::with.default(raw_data_input, data.frame(master))
 4. base::eval(substitute(expr), data, enclos = parent.frame())
 5. base::eval(substitute(expr), data, enclos = parent.frame())
 6. base::data.frame(master)
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✖ | 1       0 | plot2Att
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test_plot2Att.R:22): (code run outside of `test_that()`)
Error in `data.frame(master)`: object 'master' not found
Backtrace:
 1. discourseGT::prepareGraphs(raw_data_input, project_title)
      at test_plot2Att.R:22:0
 3. base::with.default(raw_data_input, data.frame(master))
 4. base::eval(substitute(expr), data, enclos = parent.frame())
 5. base::eval(substitute(expr), data, enclos = parent.frame())
 6. base::data.frame(master)
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✖ | 3      12 | prepareGraphs
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Failure (test_prepareGraphs.R:18): prepareGraphs works
all.equal(result$edge_list, edge_list[, c("source", "target")]) is not TRUE

`actual` is a character vector ('Names: 2 string mismatches')
`expected` is a logical vector (TRUE)

Failure (test_prepareGraphs.R:20): prepareGraphs works
result$weight_list not equal to edge_list$weight.
Modes: list, numeric
names for target but not for current
Attributes: < Modes: list, NULL >
Attributes: < names for target but not for current >
Attributes: < current is not list-like >
Length mismatch: comparison on first 1 components
Component 1: Numeric: lengths (9, 1) differ

Failure (test_prepareGraphs.R:28): prepareGraphs works
all.equal(result$edge_list, edge_list[, c("source", "target")]) is not TRUE

`actual` is a character vector ('Names: 2 string mismatches')
`expected` is a logical vector (TRUE)
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

══ Results ═══════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
Duration: 1.7 s

[ FAIL 7 | WARN 0 | SKIP 0 | PASS 17 ]
```
