### 2023-05-16 q1cui/discourseGT#6 @schochastics
There are two ways to approach this. Yours is ok but you should do a check if the file actually exists
```r
edgelist_raw <- function(input_file) {
  if (is.data.frame(input_file)) {
     df <- input_file
  } else if (is.character(input_file)) {
     if(isFALSE(file.exists(input_file)){
       stop("file not found")
     }
     df <- read.csv(input_file)
  } else {
     warning("Invalid input type. Please input file path as character or input data.frame directly.")
  }
}
```

For future iterations you can also think about including a generic function

```r
edgelist_raw <- function(x) {
    UseMethod("edgelist_raw", x)
}

edgelist_raw.default <- function(x) {
    stop("Invalid input type. Please input file path as character or input data.frame directly.", call. = FALSE)
}

edgelist_raw.character <- function(x){
    if(isFALSE(file.exists(x))){
        stop("file not found")
    } else{
        df <- read.csv(x)
        edgelist_raw(df)
    }
}

edgelist_raw.data.frame <- function(x){
    #rest of code
}
```

You dont need to implement this now, but I would encourage to learn about generic functions and implement it in a later version
