### 2023-05-25 d856d0e
Delete discourseGT.R
replace by discourseGT-package.R
