+++ vignettes/discourseGT_new.Rmd
@@ -1,289 +1,0 @@
----
-title: '\pkg{discourseGT}: An \proglang{R} package to analyze discourse networks in educational contexts'
-output: 
-  pdf_document:
-    keep_tex: true
-    number_sections: true
-    latex_engine: xelatex
-    citation_package: natbib
-fontsize: 12pt
-vignette: >
-  %\VignetteIndexEntry{discourseGT: An R package to analyze discourse networks in educational contexts}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
-header-includes:
-- |
-  ```{=latex}
-  \usepackage{hyperref}
-  \newcommand{\pkg}[1]{{\fontseries{m}\fontseries{b}\selectfont #1}}
-  %%\let\code=\texttt
-  \let\proglang=\textsf
-  \usepackage{floatrow}
-  \floatsetup[table]{capposition=top}
-  \floatsetup[figure]{capposition=top}
-  \usepackage{longtable}
-  \usepackage{booktabs}
-  \usepackage{tabulary}
-  \usepackage{ragged2e}
-  ```
-bibliography: references.bib
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-```
-
-# discourseGT Workflow {#workflow}
-## General Workflow
-The functions of \pkg{discourseGT} were designed to be as modular as possible, making it possible to only run analyses of interest. Figure \ref{fig:workflow} represents the general workflow of \pkg{discourseGT}, and Table \ref{table:dgt_func} describes explicit function names organized by their general uses.
-
-\begin{figure}[H]
-\caption{General workflow of \pkg{discourseGT}. The raw data can either be converted to an \pkg{igraph} object for further analysis or directly passed for NGT analysis. All console output can be permanently stored to the user's local disk. Green represents the start of the workflow. Purple represents steps necessary to generate an \pkg{igraph} object. Blue represents the potential downstream uses of an \pkg{igraph} object. Orange represents NGT analysis. Red signals the end of the workflow.}
-\label{fig:workflow}
-\centering
-\includegraphics[width=\textwidth]{maintext_files/figure-latex/workflow.png}
-\end{figure}
-
-
-\begin{longtable}[H]{@{}p{0.07\textwidth}p{0.27\textwidth}p{0.28\textwidth}p{0.28\textwidth}}
-\caption{List of all \pkg{discourseGT} functions}    %%%%<===
-\label{table:dgt_func}\\
-\toprule
-\textbf{Phase} & \textbf{Function Name} & \textbf{Parameter(s)} & \textbf{Description} \\ \midrule
-\endfirsthead
-\hline
-\textbf{Phase} & \textbf{Function Name} & \textbf{Parameter(s)} & \textbf{Description} \\ \midrule
-\endhead
-Purple\newline        &       \texttt{tabulate\_edges()}\newline        &       \texttt{input} – \texttt{data.frame} or \texttt{string}. Points to \texttt{.csv} file with talk-turn data in the question-and-response format.\newline\newline\texttt{iscsvfile} – \texttt{boolean}. \texttt{TRUE} if input is a \texttt{.csv} file. Else \texttt{FALSE}.\newline\newline\texttt{silentNodes} – \texttt{integer}. The number of nodes that do not interact with others.\newline       &        Calculates the weighted edge list from the input data and number of silent nodes not captured in the data.\newline \\
-
[…]
