### 2023-05-25 36b6d03
Delete test_basicPlot.R
replace by test-basicPlot.R
