### 2023-05-25 9408e53
Delete discourseGT.Rd
replaced by discourseGT-package.Rd
