+++ man/discourseGT.Rd
@@ -1,9 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/discourseGT.R
-\docType{package}
-\name{discourseGT}
-\alias{discourseGT}
-\title{discourseGT: A package to analyze groups patterns using graph theory in educational settings}
-\description{
-discourseGT is a library that utilizes the various network libraries to produce network analysis in educational settings
-}
