### 2023-05-25 #5143 @q1cui
@schochastics We would like to extend our gratitude to you for taking the time to review our manuscript and for your valuable feedback which has helped us improve the package. Herein, we would like to provide a detailed response to your concerns.
**Unit Tests**: We appreciate your suggestion to incorporate unit tests into our package. Based on your feedback, we added several tests which we found to be crucial for assessing the package’s functionality and reliability. However, we acknowledge that there were some issues initially causing a few tests to fail. After a thorough investigation, we have addressed these issues and made sure that all tests are passing.
**Generic Functions**: Your suggestion to include generic functions was insightful and has indeed improved the design of our package. We have now incorporated a series of generic functions.
**Vignette Errors**: We are thankful for your observation about the errors occurring during the building of the vignette. Upon your suggestion, we revisited our vignette build process and rectified the issues causing the errors. We have now ensured that the vignette builds successfully without any errors.
