### 2023-05-26 #5143 @zoometh
@marcosvital and @q1cui
* I suggest to change the chapter named "**Run Graph Theory Analysis**" to "**Run Graph Analysis**" or "**Run Network Analysis**" since graphs are mathematical structures, whereas networks are the objects of study.

For the rest, I agree with @schochastics
