### 2023-06-01 q1cui/discourseGT#8 @schochastics
check issues
There are some issues remaining when running `devtools::check()`
```
devtools::check(document = FALSE)

══ Building ══════════════════════════════════════════════════════════════════════════════════════
Setting env vars:
• CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
• CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
• CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX14FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX17FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX20FLAGS: -Wall -pedantic -fdiagnostics-color=always
── R CMD build ───────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/home/david/Documents/github_projects/discourseGT/DESCRIPTION’ ...
─  preparing ‘discourseGT’:
✔  creating vignettes (12.5s)
   WARNING: Removing directory ‘discourseGT/Meta’ which should only occur
     in an installed package
─  building ‘discourseGT_1.1.8.tar.gz’

══ Checking ══════════════════════════════════════════════════════════════════════════════════════
Setting env vars:
• _R_CHECK_CRAN_INCOMING_USE_ASPELL_           : TRUE
• _R_CHECK_CRAN_INCOMING_REMOTE_               : FALSE
• _R_CHECK_CRAN_INCOMING_                      : FALSE
• _R_CHECK_FORCE_SUGGESTS_                     : FALSE
• _R_CHECK_PACKAGES_USED_IGNORE_UNUSED_IMPORTS_: FALSE
• NOT_CRAN                                     : true
── R CMD check ───────────────────────────────────────────────────────────────────────────────────
─  using log directory ‘/home/david/Documents/github_projects/discourseGT.Rcheck’
─  using R version 4.2.2 Patched (2022-11-10 r83330)
─  using platform: x86_64-pc-linux-gnu (64-bit)
─  using session charset: UTF-8
─  using options ‘--no-manual --as-cran’
✔  checking for file ‘discourseGT/DESCRIPTION’
─  checking extension type ... Package
─  this is package ‘discourseGT’ version ‘1.1.8’
─  package encoding: UTF-8
✔  checking package namespace information
✔  checking package dependencies (6.3s)
✔  checking if this is a source package ...
✔  checking if there is a namespace
✔  checking for executable files ...
✔  checking for hidden files and directories ...
✔  checking for portable file names
✔  checking for sufficient/correct file permissions
✔  checking whether package ‘discourseGT’ can be installed (1.7s)
✔  checking installed package size ...
✔  checking package directory
✔  checking for future file timestamps ...
✔  checking ‘build’ directory ...
N  checking top-level files ...
   Non-standard files/directories found at top level:
     ‘LICENSE.md’ ‘doc’ ‘paper’
✔  checking for left-over files
✔  checking index information ...
✔  checking package subdirectories ...
✔  checking R files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded ...
✔  checking whether the package can be loaded with stated dependencies ...
✔  checking whether the package can be unloaded cleanly ...
✔  checking whether the namespace can be loaded with stated dependencies ...
✔  checking whether the namespace can be unloaded cleanly ...
✔  checking loading without being on the library search path ...
✔  checking dependencies in R code (776ms)
✔  checking S3 generic/method consistency (773ms)
✔  checking replacement functions ...
✔  checking foreign function calls ...
N  checking R code for possible problems (3.4s)
   plotNGTData.data.frame: no visible binding for global variable
     ‘episode_length’
   plotNGTData.data.frame: no visible binding for global variable ‘Freq’
   plotNGTData.data.frame: no visible binding for global variable
     ‘ep_starts_hour’
   plotNGTData.data.frame: no visible binding for global variable
     ‘ep_conts_hour’
   plotNGTData.data.frame: no visible binding for global variable
     ‘participant’
   plotNGTData.data.frame: no visible binding for global variable
     ‘normalized_turn_ratio’
   plotNGTData.data.frame: no visible binding for global variable ‘group’
     Freq ep_conts_hour ep_starts_hour episode_length group
     normalized_turn_ratio participant
✔  checking Rd files (380ms)
✔  checking Rd metadata ...
✔  checking Rd line widths ...
✔  checking Rd cross-references ...
✔  checking for missing documentation entries ...
✔  checking for code/documentation mismatches (537ms)
✔  checking Rd \usage sections (851ms)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples ...
✔  checking contents of ‘data’ directory ...
✔  checking data for non-ASCII characters ...
✔  checking LazyData
✔  checking data for ASCII and uncompressed saves ...
✔  checking sizes of PDF files under ‘inst/doc’ (2.5s)
✔  checking installed files from ‘inst/doc’ ...
✔  checking files in ‘vignettes’ ...
✔  checking examples (5.7s)
✔  checking for unstated dependencies in ‘tests’ ...
─  checking tests ...
✔  Running ‘testthat.R’ (2.3s)
✔  checking for unstated dependencies in vignettes ...
✔  checking package vignettes in ‘inst/doc’ ...
─  checking re-building of vignette outputs ... [14s/14s] OK (13.7s)
✔  checking for non-standard things in the check directory
✔  checking for detritus in the temp directory

   See
     ‘/home/david/Documents/github_projects/discourseGT.Rcheck/00check.log’
   for details.

── R CMD check results ──────────────────────────────────────────────────── discourseGT 1.1.8 ────
Duration: 43.9s

❯ checking top-level files ... NOTE
  Non-standard files/directories found at top level:
    ‘LICENSE.md’ ‘doc’ ‘paper’

❯ checking R code for possible problems ... NOTE
  plotNGTData.data.frame: no visible binding for global variable
    ‘episode_length’
  plotNGTData.data.frame: no visible binding for global variable ‘Freq’
  plotNGTData.data.frame: no visible binding for global variable
    ‘ep_starts_hour’
  plotNGTData.data.frame: no visible binding for global variable
    ‘ep_conts_hour’
  plotNGTData.data.frame: no visible binding for global variable
    ‘participant’
  plotNGTData.data.frame: no visible binding for global variable
    ‘normalized_turn_ratio’
  plotNGTData.data.frame: no visible binding for global variable ‘group’
    Freq ep_conts_hour ep_starts_hour episode_length group
    normalized_turn_ratio participant

0 errors ✔ | 0 warnings ✔ | 2 notes ✖

R CMD check succeeded
```

The issues are:
```
 WARNING: Removing directory ‘discourseGT/Meta’ which should only occur
     in an installed package
```

```
plotNGTData.data.frame: no visible binding for global variable
     ‘episode_length’
   plotNGTData.data.frame: no visible binding for global variable ‘Freq’
   plotNGTData.data.frame: no visible binding for global variable
     ‘ep_starts_hour’
   plotNGTData.data.frame: no visible binding for global variable
     ‘ep_conts_hour’
   plotNGTData.data.frame: no visible binding for global variable
     ‘participant’
   plotNGTData.data.frame: no visible binding for global variable
     ‘normalized_turn_ratio’
   plotNGTData.data.frame: no visible binding for global variable ‘group’
     Freq ep_conts_hour ep_starts_hour episode_length group
     normalized_turn_ratio participant
```

```
 Non-standard files/directories found at top level:
     ‘LICENSE.md’ ‘doc’ ‘paper’
```
to fix the last one, create a `.Rbuildignore` file and add the three files/dirs in a separate line each

This is part of the review for JOSS https://github.com/openjournals/joss-reviews/issues/5143
