### 2023-06-01 bde91e0
cleared notes
Notes in devtools::check() for "global variables" are not cleared.
Meta has been deleted.
New MD5 hash being generated.
