+++ R/plotNGTData.R
@@ -38,31 +38,31 @@ plotNGTData.character  <- function(data, convoMinutes, silentNodes = 0) {
 
 #' @export
 plotNGTData.data.frame <- function(data, convoMinutes, silentNodes = 0){
-
+  
   #Create a copy of raw file for processing
   raw <- data
-
+  
   #colnames(raw)[colnames(raw)=="ï..q"] <- "q"
-
+  
   ##### Histogram for Frequency of Episodes #####
   #Initalize empty columns
   #raw$select <- NA
   raw$counter <- NA
   raw$round <- NA
-
+  
   #Retrieve total number of rows
   totalrows <- nrow(raw)
-
+  
   #Classify which rows are start of new episode
   #11 is new episode defined by new question
   #99 is successive responses to question
   #raw$select[is.na(raw$q) == FALSE] <- 11
   #raw$select[is.na(raw$r) == FALSE] <- 99
-
+  
   #Initalize necessary variables outside of loop
   #Episode number counter
   q <- 0
-
+  
   #Loop through data and count number of episodes(q) and episode index number per each episode (n)
   for (i in 1:totalrows){
     if(is.na(raw$ep_start[i])==FALSE & is.na(raw$ep_cont[i]) == TRUE){
@@ -75,17 +75,17 @@ plotNGTData.data.frame <- function(data, convoMinutes, silentNodes = 0){
       n <- n + 1
       raw$counter[i] <- n
       raw$round[i] <- q
-
+      
     }
   }
-
+  
   #Collect the number of episodes
   eps_list <- unique(raw$round)
   eps <- max(eps_list)
-
+  
   #Initalize List
   eps_len_list <- NULL
-
+  
   #Map each episode number with the length of each individual episodes
   for(var in eps_list){
     #Create a temporary data frame
@@ -94,30 +94,30 @@ plotNGTData.data.frame <- function(data, convoMinutes, silentNodes = 0){
     maxCount <- max(temp$counter)
     eps_len_list <- c(eps_len_list, maxCount)
   }
-
+  
   #Create a table with the episode number and max number of occurances
   eps_master <- data.frame(eps_list, eps_len_list)
-
+  
   #Count number of occurances that each length appears and store as data frame
   eps_st <- data.frame(table(eps_master$eps_len_list))
   colnames(eps_st)[colnames(eps_st)=="Var1"] <- "episode_length"
-
+  
   # modify the df so that there is a continuous x-axis
   max <- as.integer(as.vector(eps_st$episode_length[length(eps_st$episode_length)]))
   missingNodes <- data.frame("episode_length" = factor(as.vector(1:max)), "Freq" = as.vector(as.integer(0)))
-
+  
   eps_st <- dplyr::bind_rows(eps_st, missingNodes)
   eps_st <- aggregate(Freq~episode_length, data = eps_st, FUN = sum)
   eps_st <- eps_st[order(as.integer(eps_st$episode_length)),]
   row.names(eps_st) <- 1:nrow(eps_st)
-
+  
   #Create a histogram plot of lengths
   #View all color fill options through this command: RColorBrewer::display.brewer.all()
   #Color can be an option for the function, else go with the default of blue
[…]
