### 2023-06-01 #5143 @q1cui
Thank you @schochastics for your detailed feedback and suggestions on our manuscript. We have carefully considered your comments and have made the following corrections:
We have removed the directory 'discourseGT/Meta' as per your suggestion.
We have addressed the issue of no visible binding for global variables in the 'plotNGTData.data.frame' function.
Lastly, we have created a '.Rbuildignore' file and added 'LICENSE.md', 'doc', and 'paper' directories.
We believe that these corrections have improved the quality of our work and have addressed your concerns.
