+++ DESCRIPTION
@@ -3,7 +3,6 @@ Type: Package
 Title: Analyze Group Patterns using Graph Theory in Educational
         Settings
 Version: 1.1.8
-Date: 2022-10-27
 Authors@R: c(
   person("Albert","Chai",   role=c("aut"), email = "albertchai@ucsd.edu"),
   person("Andrew","Lee",    role=("aut"), email = "aleet383@ucla.edu"),
@@ -12,8 +11,8 @@ Authors@R: c(
   person("Kevin","Banh",    role=c("ctb"), email = "kebanh@ucsd.edu"),
   person("Priya","Pahal",   role=c("ctb"), email = "ppahal@ucsd.edu"),
   person("Jitarth","Sheth", role=c("aut"), email = "jhsheth@ucsd.edu"),
-  person("Qi","Cui",        role=c("aut"), email = "q1cui@ucsd.edu"),
-  person("Stanley","Lo",    role=c("aut", "cre"), email = "smlo@ucsd.edu"))
+  person("Qi","Cui",        role=c("aut", "cre"), email = "q1cui@ucsd.edu"),
+  person("Stanley","Lo",    role=c("aut"), email = "smlo@ucsd.edu"))
 Description: Analyzes group patterns using discourse analysis data with graph theory mathematics. Takes the order of which individuals talk and converts it to a network edge and weight list. Returns the density, centrality, centralization, and subgroup information for each group. Based on the analytical framework laid out in Chai et al. (2019) <doi:10.1187/cbe.18-11-0222>.
 License: MIT + file LICENSE
 LazyData: true
@@ -42,20 +41,5 @@ Imports:
     graphics,
     stats,
     utils
-NeedsCompilation: no
-Packaged: 2022-10-27 22:59:18 UTC; jitar
-Author: 
-  Albert Chai [aut],
-  Andrew Lee [aut],
-  Joshua Le [aut],
-  Katherine Ly [ctb],
-  Kevin Banh [ctb],
-  Priya Pahal [ctb],
-  Jitarth Sheth [aut],
-  Qi Cui [aut],
-  Stanley Lo [aut, cre]
-Maintainer: Stanley Lo <smlo@ucsd.edu>
 VignetteBuilder: knitr
-Repository: CRAN
-Date/Publication: 2022-10-27 23:25:10 UTC
 Config/testthat/edition: 3
