### 2023-06-12 #5143 @marcosvital
> @marcosvital I am grateful that the paper has been accepted by two reviewers. However, I have noticed that the status of the paper has not changed on the submission portal, and I would like to confirm if there are any further steps that I need to undertake at my end. I understand that the administrative process may take some time, and I do not mean to rush. I only wish to ensure that no further actions are required from my side that could potentially delay the process. I look forward to hearing from you at your earliest convenience. Thanks!

Hi, @q1cui! There are only a few steps ahead. Everything seems to be fine with the references, so we should be able to proceed soon, ok?
