### 2023-06-12 #5143 @marcosvital
Hi, @q1cui , I believe we are almost ready to go here. For now, I would like you to:

* make a final check the latest version of the manuscript generated by Whedon, to see if everything is ok.
* archive the last release of the package (on Zenodo, figshare, or other) if you already didn't do this. After that, just let me know the archive DOI.
* let me know if 1.1.8 is still the latest release or if you have a new version number.
