### 2023-06-13 #5143 @q1cui
@marcosvital I appreciate your guidance throughout this process.
I will conduct a final check of the latest version of the manuscript generated by Whedon to ensure that everything is in order.
In regard to the software package, we are planning an update soon. The current version is 1.1.8, but we will release version 1.2.0 shortly.
Thank you for your support during this final stage of the submission process.
