### 2023-06-28 c5e9454
update_cran-comments
Perfermed checks:
urlchecker::url_check()
devtools::check(remote = TRUE, manual = TRUE)
devtools::check_win_devel()
revdepcheck::revdep_check(num_workers = 4)
