### 2023-07-08 e7cee17
update vignette docs
switch vignette builder to R.rsp
use asis as vignette engine for pre-built pdf vignette
removed doc file
