### 2023-07-08 77fe6a6
initial readme.rmd
switch to readme.rmd to generate readme file
