### 2023-07-12 45e9218
update cran_comments for submission
Added r-hub, win_devel, local r cmd check results.
