### 2023-07-18 #5143 @q1cui
@marcosvital I'm writing to provide an update on our project. We submitted an updated version (1.2.0) of our package to CRAN. This revision includes several improvements, most of which are based on the insightful comments received from our reviewers. Currently, we are waiting for confirmation from CRAN about the acceptance of our updated package.

At the same time, I'd like to discuss the status of our manuscript tied to this package. As the manuscript has been updated in line with the reviewers' comments and the package is also revised accordingly, I was wondering if the status of our manuscript could now be changed to 'Accepted'. Alternatively, should we wait for the formal approval of the package update from CRAN before any status change?

Your guidance on this matter would be extremely helpful to us as we navigate these next steps. If you require additional details or if there is anything else you need in the meantime, please feel free to let us know.
