### 2023-07-19 #5143 @q1cui
@marcosvital Following your request, I have thoroughly examined the latest version of the manuscript and I am pleased to confirm that everything appears to be in order.
Regarding the software package, we have indeed made some revisions based on the valuable feedback from the reviewers, and I'm happy to report that the updated version, 1.2.0, is now available on CRAN.
I have also archived this latest release on Zenodo. You can access it using the following DOI: 10.5281/zenodo.8164950.
I appreciate your time and efforts in guiding us through this process. Should you require any further information or clarifications, please do not hesitate to ask.
