### 2023-08-02 3c13ea5
Update references.bib
delete duplicate reference
