### 2023-08-02 #5143 @q1cui
> ⚠️ [Error preparing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/5742808972?check_suite_focus=true). The generated XML metadata file is invalid.
>
> ```yaml
> ID ref-Lou_2001 already defined
> ```
We have deleted the duplicate reference.
