### 2023-08-02 98dcdae
Update paper.md
delete the figure
