### 2023-08-02 #5143 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
